+++ lib/bootstrap_form/form_builder.rb
@@ -51,11 +51,12 @@ module BootstrapForm
       end
     end
 
-    def check_box(name, options = {}, checked_value = '1', unchecked_value = '0')
+    def check_box(name, options = {}, checked_value = '1', unchecked_value = '0', &block)
       options = options.symbolize_keys!
 
       html = super(name, options.except(:label, :help, :inline), checked_value, unchecked_value)
-      html.concat(' ').concat(options[:label] || object.class.human_attribute_name(name) || name.to_s.humanize)
+      label_content = block_given? ? capture(&block) : options[:label]
+      html.concat(' ').concat(label_content || object.class.human_attribute_name(name) || name.to_s.humanize)
 
       if options[:inline]
         label(name, html, class: "checkbox-inline")